requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pdfplumber
pymongo
python-multipart
python-jose
PyJWT
google-generativeai